Enable basic linter checks
This surfaces some unreachable code, unused functions and a lot of
unchecked errors.

The housekeep function in particular tripped me up when I was debugging.
It 'fails' with a `redis.Nil` response. It's not clear if that's an
issue or just how the Lua scripts are responding.

We can add [more linters](https://golangci-lint.run/usage/linters) of course; this is just the defaults. Some
useful ones to look at:
 - contextcheck
 - gochecknoglobals
 - gofmt/gofumpt
 - goimports