Add GetRoomState() facility to API client
An operator may want to examine the state of a room in case something
goes terribly wrong, to see if there are clients which are stuck for
an abnormally long time, and unsubscribe them from the room manually
if this is the case.

To support this use case, we add a `GetRoomState()` function to
`RedisQueueApiClient` interface which returns a `RoomState` structure,
carrying Room ID, Room Message Backlog length, a map of known clients
and their last activity timestamp, and a map of acknowledged clients
along with their ACK timestamp.

This information should be enough to make a decision on whether one or
more clients should be forcefully unsubscribed from the room.